♻️ refactor: colocate boxsh client with plugin
