+++ plugins/sandbox/boxsh/boxshclient/coretools_helpers_test.go
